+++ packages/fx-core/src/core/index.ts
@@ -53,6 +53,7 @@ import {
   deserializeDict,
   fetchCodeZip,
   getStrings,
+  isValidProject,
   mapToJson,
   mergeSerectData,
   objectToMap,
@@ -75,6 +76,7 @@ import {
 import * as jsonschema from "jsonschema";
 import { sleep } from "../plugins/resource/spfx/utils/utils";
 import AdmZip from "adm-zip";
+export * from "./error";
 
 class CoreImpl implements Core {
   private target?: CoreImpl;
@@ -443,7 +445,7 @@ class CoreImpl implements Core {
       supported = false;
     } else {
       this.ctx.root = workspace;
-      supported = await this.isSupported();
+      supported = isValidProject(workspace);
       if (!supported) {
         this.ctx.logProvider?.warning(`non Teams project:${workspace}`);
       } else {
@@ -811,24 +813,6 @@ class CoreImpl implements Core {
     }
   }
 
-  public async isSupported(workspace?: string): Promise<boolean> {
-    let p = this.ctx.root;
-    if (workspace) {
-      p = workspace;
-    }
-    // some validation
-    const checklist: string[] = [
-      `${p}/.${ConfigFolderName}/settings.json`,
-      `${p}/.${ConfigFolderName}/env.default.json`,
-    ];
-    for (const fp of checklist) {
-      if (!(await fs.pathExists(path.resolve(fp)))) {
-        return false;
-      }
-    }
-    return true;
-  }
-
   public async readConfigs(): Promise<Result<null, FxError>> {
     if (!fs.existsSync(`${this.ctx.root}/.${ConfigFolderName}`)) {
       this.ctx.logProvider?.warning(
@@ -1063,7 +1047,7 @@ class CoreImpl implements Core {
             description: "",
             author: "",
             scripts: {
-              test: "echo \"Error: no test specified\" && exit 1",
+              test: 'echo "Error: no test specified" && exit 1',
             },
             license: "MIT",
           },
@@ -1156,7 +1140,7 @@ export class CoreProxy implements Core {
     try {
       // check if this project is supported
       if (checkAndConfig) {
-        const supported = await this.coreImpl.isSupported();
+        const supported = isValidProject(this.coreImpl.ctx.root);
         if (!supported) {
           return notSupportedRes;
         }
@@ -1244,7 +1228,7 @@ export class CoreProxy implements Core {
     return await this.runWithErrorHandling<QTreeNode | undefined>(
       "getQuestionsForUserTask",
       true,
-      err(error.NotSupportedProjectType()),
+      err(error.InvalidProjectError),
       () => this.coreImpl.getQuestionsForUserTask(func, platform)
     );
   }
@@ -1258,7 +1242,7 @@ export class CoreProxy implements Core {
     return await this.runWithErrorHandling<QTreeNode | undefined>(
       "executeUserTask",
       check,
-      err(error.NotSupportedProjectType()),
+      err(error.InvalidProjectError),
       () => this.coreImpl.executeUserTask(func, answers),
       answers
     );
@@ -1310,7 +1294,7 @@ export class CoreProxy implements Core {
     return await this.runWithErrorHandling<null>(
       "localDebug",
       true,
-      err(error.NotSupportedProjectType()),
+      err(error.InvalidProjectError),
       () => this.coreImpl.localDebug(answers),
       answers
     );
@@ -1319,7 +1303,7 @@ export class CoreProxy implements Core {
     return await this.runWithErrorHandling<null>(
       "provision",
       true,
-      err(error.NotSupportedProjectType()),
+      err(error.InvalidProjectError),
       () => this.coreImpl.provision(answers),
       answers
     );
@@ -1328,7 +1312,7 @@ export class CoreProxy implements Core {
     return await this.runWithErrorHandling<null>(
       "deploy",
       true,
-      err(error.NotSupportedProjectType()),
+      err(error.InvalidProjectError),
       () => this.coreImpl.deploy(answers),
       answers
     );
@@ -1337,7 +1321,7 @@ export class CoreProxy implements Core {
     return await this.runWithErrorHandling<null>(
       "publish",
       true,
-      err(error.NotSupportedProjectType()),
+      err(error.InvalidProjectError),
       () => this.coreImpl.publish(answers),
       answers
     );
@@ -1346,7 +1330,7 @@ export class CoreProxy implements Core {
     return await this.runWithErrorHandling<null>(
       "createEnv",
       true,
-      err(error.NotSupportedProjectType()),
+      err(error.InvalidProjectError),
       () => this.coreImpl.createEnv(env)
     );
   }
@@ -1354,7 +1338,7 @@ export class CoreProxy implements Core {
     return await this.runWithErrorHandling<null>(
       "removeEnv",
       true,
-      err(error.NotSupportedProjectType()),
+      err(error.InvalidProjectError),
       () => this.coreImpl.removeEnv(env)
     );
   }
@@ -1362,7 +1346,7 @@ export class CoreProxy implements Core {
     return await this.runWithErrorHandling<null>(
       "switchEnv",
       true,
-      err(error.NotSupportedProjectType()),
+      err(error.InvalidProjectError),
       () => this.coreImpl.switchEnv(env)
     );
   }
@@ -1370,7 +1354,7 @@ export class CoreProxy implements Core {
     return await this.runWithErrorHandling<string[]>(
       "listEnvs",
       true,
-      err(error.NotSupportedProjectType()),
+      err(error.InvalidProjectError),
       () => this.coreImpl.listEnvs()
     );
   }
