+++ packages/vscode-extension/src/error.ts
@@ -15,15 +15,3 @@ export enum ExtensionErrors {
   UnknownSubscription = "UnknownSubscription",
   PortAlreadyInUse = "PortAlreadyInUse"
 }
-
-export const NoProjectOpenedError = new UserError(
-  "NoProjectOpened",
-  "No project opened, you can create a new project or open an existing one.",
-  ExtensionSource
-);
-
-export const InvalidProject = new UserError(
-  "InvalidProject",
-  "The project type is invalid",
-  ExtensionSource
-);
